requirements.txt
flask==2.0.1
werkzeug==2.0.3
pillow==9.0.1
colorthief==0.2.1
numpy==1.22.3